% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmalizr-package.R
\docType{package}
\name{flowmalizr-package}
\alias{flowmalizr-package}
\alias{_PACKAGE}
\alias{documentation}
\title{flowmalizr: Normalizing Cell Population Percentages}
\description{
This needs a description.
}
\author{
\strong{Maintainer}: Pablo Maldonado \email{mal.j.pablo@gmail.com}

Authors:
\itemize{
  \item Marcela Henao-Tamayo \email{marcela.henao_tamayo@colostate.edu}
}

}
\keyword{internal}
