% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmalizr.R
\name{flowmalizr}
\alias{flowmalizr}
\title{make_df}
\usage{
flowmalizr(path)
}
\arguments{
\item{path}{}
}
\value{
tidy dataframe made from xls input
}
\description{
make_df
}
\examples{
path_to_data <- system.file("extdata", "example_data.xlsx", package = "flowmalizr")
flowmalizr(path = path_to_data)
}
