% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmalizr.R
\name{group_v_group}
\alias{group_v_group}
\title{1v1 comparison}
\usage{
group_v_group(groupA, groupB)
}
\arguments{
\item{groupA}{is group}

\item{groupB}{is group}
}
\value{
A 1v1 visual of phenotype per group
}
\description{
\code{roup_v_group()} allows for visual group vs group comparison of analyzed phenotypes
}
\examples{
group_v_group(1, 4)
}
