% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmalizr.R
\name{sep_groups}
\alias{sep_groups}
\title{Separate group and replicate into their own columns}
\usage{
sep_groups()
}
\value{
New df with group and replicate in their own columns
}
\description{
\code{sep_group()} takes the newly generated \code{imported_df} and separates
groups from replicates for more accurate analysis when determining phenotypes
observed per group
}
\examples{
sep_groups()

}
