% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmalizr.R
\name{unique_pops}
\alias{unique_pops}
\title{Pull unique gated populations}
\usage{
unique_pops()
}
\value{

}
\description{
\code{unique_pops()} provides a quick view of all phenotypes to be analyzed using this package
}
\examples{
unique_pops()

}
