% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmalizr.R
\name{visualize_groups}
\alias{visualize_groups}
\title{Plot percentage of unique populations}
\usage{
visualize_groups()
}
\value{

}
\description{
\code{visualize_groups()} plots and averages biological replicates for visualization of
all groups and phenotypes
}
\examples{
visualize_groups()
}
