% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmalizr.R
\name{flowmalizr}
\alias{flowmalizr}
\title{Normalize cell counts and compare cell phenotype}
\usage{
flowmalizr(path)
}
\arguments{
\item{groupB.}{}
}
\value{
tidy dataframe made from xls input
}
\description{
Normalize cell counts and compare cell phenotype
}
\examples{
path_to_data <- system.file("extdata", "example_data.xlsx", package = "flowmalizr")
flowmalizr(path = path_to_data)
}
