% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmalizr.R
\name{flowmalizr}
\alias{flowmalizr}
\title{Normalize flow data cell counts and compare cell phenotypes}
\usage{
flowmalizr(path)
}
\arguments{
\item{path}{chracter string boah blah path_to_data}
}
\value{
Longer dataframe with data under new "name", "cells_from_total", and
"percentage_of_total" columns.
}
\description{
\code{flowmalizr()} takes an xls input sheet containing columns listed as
groups, total_cell_count_per_mL, live_cells, and phenotypes of interest.
Each row contains a group-replicate, cell number per mL, live cell count,
and cells identified, respectively. This is used to detemine
cells_from_total and percentage_of_total by
(total_cell_count_per_mL*cells-of-phenotype/live_cells)*100.
The output from this function is a dataframe that can be manipulated for
visualization of cell phenotypes from each group
}
\examples{
path_to_data <- system.file("extdata", "example_data.xlsx", package = "flowmalizr")
flowmalizr(path_to_data)

}
