% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmalizr.R
\name{group_v_group}
\alias{group_v_group}
\title{1v1 comparison}
\usage{
group_v_group(groupA, groupB)
}
\arguments{
\item{x}{groupA}

\item{y}{groupB}
}
\value{

}
\description{
\code{roup_v_group()} allows for visual group vs group comparison of analyzed phenotypes
}
\examples{
group_v_group(1, 4)
}
